Update ts1.py
Probably something with line 33 and down from there. I printed the dictionary so you can see it does it correctly, issue was happening when the client had to receive the 2nd line of bits. So when there was only one line it would work but any more and the client had issues.